fix(cli): three Windows CI failures in the OS-sensitive go-test leg
The OPS-21 Win/mac go-test leg surfaced three GOOS-specific bugs:

1. skillgen (14 test failures): the embedded content/*.md are go:embed'd,
   and a Windows checkout (core.autocrlf) embeds CRLF bytes, so
   splitFrontmatter's `---\n` fence match failed and every skill parsed as
   "missing name in frontmatter". Fixed defensively (normalize CRLF/CR->LF
   in splitFrontmatter, since the backend serves these bytes too) and
   structurally (.gitattributes eol=lf on both content mirrors + the source
   SKILL.md, so the embedded bytes are deterministic on every platform).

2. wireGracefulCancel used syscall.SIGTERM, which os.Process.Signal rejects
   on Windows ("not supported by windows"). Split the terminate primitive by
   GOOS (run_cancel_unix.go: SIGTERM so sudo can relay it down the launch
   chain; run_cancel_windows.go: Kill), mirroring the OPS-20 proc split. The
   SIGTERM test moves to a !windows file; a Windows counterpart asserts the
   Kill-on-cancel contract.

3. logs_test hard-coded a forward-slash path; resolveLogPath correctly
   filepath.Join's the configured POSIX path to the host separator, so the
   assertion now compares against filepath.FromSlash.
