+++ .gitattributes
@@ -3,5 +3,13 @@
 # `make skills` (tools/skills_sync.py) and pinned by tests/arch/test_skill_drift.py.
 # Do not hand-edit — edit the source and regenerate. `linguist-generated`
 # collapses them in GitHub diffs and keeps them out of language stats.
-cli/internal/skillgen/content/*.md linguist-generated=true
-src/jentic_one/shared/web/content/*.md linguist-generated=true
+#
+# `eol=lf` pins these to LF in the working tree on every platform. The CLI
+# `go:embed`s cli/internal/skillgen/content/*.md into the binary, so a Windows
+# checkout converting them to CRLF (core.autocrlf) would embed CRLF bytes and
+# break the frontmatter parser (splitFrontmatter also normalizes defensively).
+# The source SKILL.md files are pinned too so the mirrors stay byte-identical.
+cli/internal/skillgen/content/*.md linguist-generated=true eol=lf
+src/jentic_one/shared/web/content/*.md linguist-generated=true eol=lf
+skills/**/SKILL.md eol=lf
+
