+++ cli/internal/cli/localagentcmd/run_test.go
@@ -2,11 +2,8 @@ package localagentcmd
 
 import (
 	"bytes"
-	"context"
 	"errors"
-	"os/exec"
 	"strings"
-	"syscall"
 	"testing"
 
 	"github.com/jentic/jentic-one/cli/internal/cli/ux"
@@ -214,37 +211,10 @@ func TestWarnSameUserOnce(t *testing.T) {
 	}
 }
 
-// TestWireGracefulCancelSendsSIGTERM proves cancellation terminates the child with
-// a catchable SIGTERM (which sudo can relay down the launch chain) rather than the
-// exec default SIGKILL to the direct child, and sets a WaitDelay SIGKILL backstop.
-func TestWireGracefulCancelSendsSIGTERM(t *testing.T) {
-	ctx, cancel := context.WithCancel(context.Background())
-	defer cancel()
-	// A long-lived child so the process is alive when we exercise Cancel.
-	c := exec.CommandContext(ctx, "sleep", "60")
-	wireGracefulCancel(c)
-	if c.WaitDelay != cancelGracePeriod {
-		t.Errorf("WaitDelay = %v, want %v", c.WaitDelay, cancelGracePeriod)
-	}
-	if c.Cancel == nil {
-		t.Fatal("Cancel must be set so ctx-cancel sends SIGTERM, not SIGKILL")
-	}
-	if err := c.Start(); err != nil {
-		t.Fatalf("start: %v", err)
-	}
-	if err := c.Cancel(); err != nil {
-		t.Fatalf("Cancel: %v", err)
-	}
-	err := c.Wait()
-	var exit *exec.ExitError
-	if !errors.As(err, &exit) {
-		t.Fatalf("expected ExitError from a signalled child, got %v", err)
-	}
-	ws, ok := exit.Sys().(syscall.WaitStatus)
-	if !ok || !ws.Signaled() || ws.Signal() != syscall.SIGTERM {
-		t.Errorf("child should have been terminated by SIGTERM, got %v", exit)
-	}
-}
+// TestWireGracefulCancelSendsSIGTERM lives in run_cancel_unix_test.go (it uses
+// POSIX signals + `sleep`); the Windows Kill()-on-cancel contract is covered by
+// run_cancel_windows_test.go. wireGracefulCancel's cross-platform wiring
+// (Cancel set, WaitDelay) is asserted in both.
 
 func TestClassifyGrantStderr(t *testing.T) {
 	cases := []struct {
